+++ R/period.R
@@ -68,9 +68,39 @@ unique.yearweek <- function(x, incomparables = FALSE, ...) {
 
 #' @export
 diff.yearweek <- function(x, lag = 1, differences = 1, ...) {
-  out <- diff((x - as_date("1969-12-29")) / 7, 
+  out <- diff((as_date(x) - as_date("1969-12-29")) / 7, 
     lag = lag, differences = differences)
   structure(out, class = "difftime", units = "weeks")
+}
+
+#' @export
+`+.yearweek` <- function(e1, e2) {
+  if (nargs() == 1L) return(e1)
+  e1_yrwk <- is_yearweek(e1)
+  e2_yrwk <- is_yearweek(e2)
+  if (e1_yrwk && e2_yrwk) {
+    abort("binary `+` is not defined for `yearweek` objects")
+  }
+  if (e1_yrwk) {
+    return(yearweek(as_date(e1) + e2 * 7))
+  } else {
+    return(yearweek(e1 * 7 + as_date(e2)))
+  }
+}
+
+#' @export
+`-.yearweek` <- function(e1, e2) {
+  if (nargs() == 1L) return(e1)
+  e1_yrwk <- is_yearweek(e1)
+  e2_yrwk <- is_yearweek(e2)
+  if (e1_yrwk && e2_yrwk) {
+    abort("binary `-` is not defined for `yearweek` objects")
+  }
+  yearweek(as_date(e1) - e2 * 7)
+}
+
+is_yearweek <- function(x) {
+  inherits(x, "yearweek")
 }
 
 #' @export
@@ -106,7 +136,8 @@ format.yearweek <- function(x, format = "%Y W%V", ...) {
   wks <- as.integer(floor((x - ord - wday + 10) / 7))
   yrs <- yr
   yrs[wks == 0] <- yr[wks == 0] - 1
-  yrs[wks == 53] <- yr[wks == 53] + !is_53weeks(yr)
+  is_53 <- yr[wks == 53]
+  yrs[wks == 53] <- is_53 + !is_53weeks(is_53)
   if (format == "%Y W%V") {
     return(paste(yrs, strftime(x, format = "W%V")))
   }
@@ -131,6 +162,7 @@ format.yearweek <- function(x, format = "%Y W%V", ...) {
 #' @examples
 #' is_53weeks(2015:2016)
 is_53weeks <- function(year) {
+  if (is_empty(year)) return(FALSE)
   if (!is_bare_numeric(year) || year < 1) {
     abort("`year` must be positive integers.")
   }
@@ -193,6 +225,36 @@ diff.yearmonth <- function(x, lag = 1, differences = 1, ...) {
   out <- diff((lubridate::year(x) - 1970) * 12 + lubridate::month(x), 
     lag = lag, differences = differences)
   structure(out, class = "difftime", units = "month")
+}
+
+#' @export
+`+.yearmonth` <- function(e1, e2) {
+  if (nargs() == 1L) return(e1)
+  e1_yrmth <- is_yearmonth(e1)
+  e2_yrmth <- is_yearmonth(e2)
+  if (e1_yrmth && e2_yrmth) {
+    abort("binary `+` is not defined for `yearmonth` objects")
+  }
+  if (e1_yrmth) {
+    return(yearmonth(as_date(e1) + lubridate::period(e2, units = "month")))
+  } else {
+    return(yearmonth(lubridate::period(e1, units = "month") + as_date(e2)))
+  }
+}
+
+#' @export
+`-.yearmonth` <- function(e1, e2) {
+  if (nargs() == 1L) return(e1)
+  e1_yrmth <- is_yearmonth(e1)
+  e2_yrmth <- is_yearmonth(e2)
+  if (e1_yrmth && e2_yrmth) {
+    abort("binary `-` is not defined for `yearmonth` objects")
+  }
+  yearmonth(as_date(e1) - lubridate::period(e2, units = "month"))
+}
+
+is_yearmonth <- function(x) {
+  inherits(x, "yearmonth")
 }
 
 #' @export
@@ -284,6 +346,36 @@ diff.yearquarter <- function(x, lag = 1, differences = 1, ...) {
   out <- diff((lubridate::year(x) - 1970) * 4 + lubridate::quarter(x),
     lag = lag, differences = differences)
   structure(out, class = "difftime", units = "quarter")
+}
+
+#' @export
+`+.yearquarter` <- function(e1, e2) {
+  if (nargs() == 1L) return(e1)
+  e1_yrqtr <- is_yearquarter(e1)
+  e2_yrqtr <- is_yearquarter(e2)
+  if (e1_yrqtr && e2_yrqtr) {
+    abort("binary `+` is not defined for `yearquarter` objects")
+  }
+  if (e1_yrqtr) {
+    return(yearquarter(as_date(e1) + lubridate::period(e2 * 3, units = "month")))
+  } else {
+    return(yearquarter(lubridate::period(e1 * 3, units = "month") + as_date(e2)))
+  }
+}
+
+#' @export
+`-.yearquarter` <- function(e1, e2) {
+  if (nargs() == 1L) return(e1)
+  e1_yrqtr <- is_yearquarter(e1)
+  e2_yrqtr <- is_yearquarter(e2)
+  if (e1_yrqtr && e2_yrqtr) {
+    abort("binary `-` is not defined for `yearquarter` objects")
+  }
+  yearquarter(as_date(e1) - lubridate::period(e2 * 3, units = "month"))
+}
+
+is_yearquarter <- function(x) {
+  inherits(x, "yearquarter")
 }
 
 #' @export
@@ -461,7 +553,7 @@ units_since <- function(x) {
 
 #' @export
 units_since.yearweek <- function(x) {
-  as.numeric((x - as_date("1969-12-29")) / 7)
+  as.numeric((as_date(x) - as_date("1969-12-29")) / 7)
 }
 
 #' @export